Rename `app/routes/search.js` to `app/search-router.js`
